Get layer attribute values tibble

Get layer attribute values tibble

Usage

layer_attributes_tbl(
  wfs = NULL,
  service = NULL,
  service_version = "2.0.0",
  layer
)

Arguments

wfs

A WFSClient R6 object with methods for interfacing an OGC Web Feature Service.

service

the EMODnet OGC WFS service name. For available services, see emodnet_wfs.

service_version

the WFS service version. Defaults to "2.0.0".

layer

character sting of layer name. To get info on layers, including layer_name use emodnet_get_wfs_info().

Value

tibble of layer attribute (variable) values with geometry column removed.

Details

Request excluding spatial information can be significantly faster. Can be useful for inspecting attribute values and constructing feature filters for more targeted and faster layer download.

Examples

layer_attributes_tbl(service = "human_activities", layer = "maritimebnds")
#> ✔ WFS client created succesfully
#> ℹ Service: 'https://ows.emodnet-humanactivities.eu/wfs'
#> ℹ Version: '2.0.0'
#> # A tibble: 1,048 × 12
#>    gml_id     objectid mblszotpid localid sitename legalfound legalfou_1 country
#>    <chr>         <dbl>      <int>   <dbl> <chr>    <date>     <chr>      <chr>  
#>  1 maritimeb…        1          7   49469 Interna… NA         NA         Germany
#>  2 maritimeb…        2          7   49471 Interna… NA         NA         Bulgar…
#>  3 maritimeb…        3          7   49472 Interna… NA         NA         Croatia
#>  4 maritimeb…        4          7   49473 Interna… NA         NA         Denmark
#>  5 maritimeb…        5          7   49474 Interna… NA         NA         Estonia
#>  6 maritimeb…        6          7   49475 Interna… NA         NA         Finland
#>  7 maritimeb…        7          7   49476 Interna… NA         NA         France 
#>  8 maritimeb…        8          7   49478 Interna… NA         NA         Ireland
#>  9 maritimeb…        9          7   49479 Interna… NA         NA         Italy  
#> 10 maritimeb…       10          7   49480 Interna… NA         NA         Latvia 
#> # … with 1,038 more rows, and 4 more variables: nationalle <chr>,
#> #   nutscode <chr>, mblsds_mbl <chr>, shape_leng <dbl>