Get EMODnet WFS layers

Performs an WFS getFeature request for layers from a wfs object or specified EMODnet Service. Filtering of layer features can also be handled via ECQL language filters.

emodnet_get_layers(
  wfs = NULL,
  service = NULL,
  service_version = "2.0.0",
  layers,
  crs = NULL,
  cql_filter = NULL,
  reduce_layers = FALSE,
  suppress_warnings = FALSE
)

Arguments

wfs
A WFSClient R6 object with methods for interfacing an OGC Web Feature Service.

service
the EMODnet OGC WFS service name. For available services, see emodnet_wfs.

service_version
the WFS service version. Defaults to "2.0.0".

layers
a character vector of layer names. To get info on layers, including layer_name use emodnet_get_wfs_info().

crs
integer. EPSG code for the output crs. If NULL (default), layers are returned with original crs.

cql_filter
character. Features returned can be filtered using valid Extended Common Query Language (ECQL) filtering statements (https://docs.geoserver.org/stable/en/user/filter/ecql_reference.html). Should be one of:

character string or character vector of length 1. Filter will be recycled across all layers requested

character vector of length equal to the length of layers. Filter will be matched to layers sequentially. Elements containing NA are ignored

named character vector. Each filter will be applied to the layer corresponding to the filter name. Filters with names that do not correspond to any layers are ignored. Layers without corresponding filters are returned whole

reduce_layers
whether to reduce output layers to a single sf object.

suppress_warnings
logical. Whether to suppress messages of layer download failures.

Value

If reduce_layers = FALSE (default), a list of sf objects, one element for each layer. Any layers for which download was unsuccessful will be NULL. If reduce_layers = TRUE, all layers are reduced (if possible) to a single sf containing data for all layers. NULL layers are ignored. reduce_layers = TRUE can also be used to return an sf out of a single layer request instead of a list of length 1.

Examples

emodnet_get_layers(service = "seabed_habitats_individual_habitat_map_and_model_datasets",
                   layers = c("dk003069", "dk003070"))

#> Loading ISO 19139 XML schemas...

#> Loading ISO 19115 codelists...

#> Loading IANA mime types...

#> No encoding supplied: defaulting to UTF-8.

#> ✔ WFS client created succesfully

#> ℹ Service: 'https://ows.emodnet-seabedhabitats.eu/geoserver/emodnet_open_maplibrary/wfs'

#> ℹ Version: '2.0.0'

#> $dk003069
#> Simple feature collection with 82 features and 8 fields
#> Geometry type: MULTISURFACE
#> Dimension:     XY
#> Bounding box:  xmin: 1065918 ymin: 7318084 xmax: 1140377 ymax: 7385447
#> Projected CRS: WGS 84 / Pseudo-Mercator
#> First 10 features:
#>            gml_id   gid      gui polygon annexi         subtype confidence
#> 1  dk003069.39813 39813 DK003069      30   1170 Geogenic origin       High
#> 2  dk003069.39847 39847 DK003069      64   1110            <NA>       High
#> 3  dk003069.39794 39794 DK003069      11   1170 Geogenic origin       High
#> 4  dk003069.39787 39787 DK003069       4   1170 Geogenic origin       High
#> 5  dk003069.39809 39809 DK003069      26   1170 Geogenic origin       High
#> 6  dk003069.39806 39806 DK003069      23   1170 Geogenic origin       High
#> 7  dk003069.39801 39801 DK003069      18   1170 Geogenic origin       High
#> 8  dk003069.39805 39805 DK003069      22   1170 Geogenic origin       High
#> 9  dk003069.39827 39827 DK003069      44   1170 Geogenic origin       High
#> 10 dk003069.39849 39849 DK003069      66   1110            <NA>       High
#>    val_comm                           geom
#> 1      <NA> MULTISURFACE (POLYGON ((113...
#> 2      <NA> MULTISURFACE (POLYGON ((112...
#> 3      <NA> MULTISURFACE (POLYGON ((112...
#> 4      <NA> MULTISURFACE (POLYGON ((109...
#> 5      <NA> MULTISURFACE (POLYGON ((113...
#> 6      <NA> MULTISURFACE (POLYGON ((110...
#> 7      <NA> MULTISURFACE (POLYGON ((109...
#> 8      <NA> MULTISURFACE (POLYGON ((109...
#> 9      <NA> MULTISURFACE (POLYGON ((112...
#> 10     <NA> MULTISURFACE (POLYGON ((111...
#> 
#> $dk003070
#> Simple feature collection with 30 features and 8 fields
#> Geometry type: MULTISURFACE
#> Dimension:     XY
#> Bounding box:  xmin: 1268645 ymin: 7276003 xmax: 1332262 ymax: 7290836
#> Projected CRS: WGS 84 / Pseudo-Mercator
#> First 10 features:
#>            gml_id   gid      gui polygon annexi         subtype confidence
#> 1  dk003070.39886 39886 DK003070      21   1170 Geogenic origin       High
#> 2  dk003070.39890 39890 DK003070      25   1170 Geogenic origin       High
#> 3  dk003070.39882 39882 DK003070      17   1170 Geogenic origin       High
#> 4  dk003070.39875 39875 DK003070      10   1170 Geogenic origin       High
#> 5  dk003070.39881 39881 DK003070      16   1170 Geogenic origin       High
#> 6  dk003070.39887 39887 DK003070      22   1170 Geogenic origin       High
#> 7  dk003070.39885 39885 DK003070      20   1170 Geogenic origin       High
#> 8  dk003070.39879 39879 DK003070      14   1170 Geogenic origin       High
#> 9  dk003070.39867 39867 DK003070       2   1170 Geogenic origin       High
#> 10 dk003070.39873 39873 DK003070       8   1170 Geogenic origin       High
#>    val_comm                           geom
#> 1      <NA> MULTISURFACE (POLYGON ((131...
#> 2      <NA> MULTISURFACE (POLYGON ((131...
#> 3      <NA> MULTISURFACE (POLYGON ((128...
#> 4      <NA> MULTISURFACE (POLYGON ((130...
#> 5      <NA> MULTISURFACE (POLYGON ((128...
#> 6      <NA> MULTISURFACE (POLYGON ((131...
#> 7      <NA> MULTISURFACE (POLYGON ((131...
#> 8      <NA> MULTISURFACE (POLYGON ((129...
#> 9      <NA> MULTISURFACE (POLYGON ((129...
#> 10     <NA> MULTISURFACE (POLYGON ((128...
#> 

emodnet_get_layers(service = "seabed_habitats_individual_habitat_map_and_model_datasets",
                   layers = c("dk003069", "dk003070"), reduce_layers = TRUE)

#> ✔ WFS client created succesfully

#> ℹ Service: 'https://ows.emodnet-seabedhabitats.eu/geoserver/emodnet_open_maplibrary/wfs'

#> ℹ Version: '2.0.0'

#> Simple feature collection with 112 features and 8 fields
#> Geometry type: MULTISURFACE
#> Dimension:     XY
#> Bounding box:  xmin: 1065918 ymin: 7276003 xmax: 1332262 ymax: 7385447
#> Projected CRS: WGS 84 / Pseudo-Mercator
#> First 10 features:
#>            gml_id   gid      gui polygon annexi         subtype confidence
#> 1  dk003069.39813 39813 DK003069      30   1170 Geogenic origin       High
#> 2  dk003069.39847 39847 DK003069      64   1110            <NA>       High
#> 3  dk003069.39794 39794 DK003069      11   1170 Geogenic origin       High
#> 4  dk003069.39787 39787 DK003069       4   1170 Geogenic origin       High
#> 5  dk003069.39809 39809 DK003069      26   1170 Geogenic origin       High
#> 6  dk003069.39806 39806 DK003069      23   1170 Geogenic origin       High
#> 7  dk003069.39801 39801 DK003069      18   1170 Geogenic origin       High
#> 8  dk003069.39805 39805 DK003069      22   1170 Geogenic origin       High
#> 9  dk003069.39827 39827 DK003069      44   1170 Geogenic origin       High
#> 10 dk003069.39849 39849 DK003069      66   1110            <NA>       High
#>    val_comm                           geom
#> 1      <NA> MULTISURFACE (POLYGON ((113...
#> 2      <NA> MULTISURFACE (POLYGON ((112...
#> 3      <NA> MULTISURFACE (POLYGON ((112...
#> 4      <NA> MULTISURFACE (POLYGON ((109...
#> 5      <NA> MULTISURFACE (POLYGON ((113...
#> 6      <NA> MULTISURFACE (POLYGON ((110...
#> 7      <NA> MULTISURFACE (POLYGON ((109...
#> 8      <NA> MULTISURFACE (POLYGON ((109...
#> 9      <NA> MULTISURFACE (POLYGON ((112...
#> 10     <NA> MULTISURFACE (POLYGON ((111...

emodnet_get_layers(service = "human_activities", layers = "maritimebnds",
                   cql_filter = "sitename='Territory sea (12 nm)'", reduce_layers = TRUE)

#> ✔ WFS client created succesfully

#> ℹ Service: 'https://ows.emodnet-humanactivities.eu/wfs'

#> ℹ Version: '2.0.0'

#> Simple feature collection with 64 features and 12 fields
#> Geometry type: MULTICURVE
#> Dimension:     XY
#> Bounding box:  xmin: -178.3897 ymin: -50.21866 xmax: 169.1762 ymax: 83.8746
#> CRS:           +init=epsg:4326
#> First 10 features:
#>             gml_id objectid mblszotpid localid              sitename legalfound
#> 1  maritimebnds.49       49          1   49024 Territory sea (12 nm)       <NA>
#> 2  maritimebnds.58       58          1   49064 Territory sea (12 nm)       <NA>
#> 3  maritimebnds.59       59          1   49065 Territory sea (12 nm)       <NA>
#> 4  maritimebnds.60       60          1   49066 Territory sea (12 nm)       <NA>
#> 5  maritimebnds.61       61          1   49087 Territory sea (12 nm)       <NA>
#> 6  maritimebnds.62       62          1   49099 Territory sea (12 nm)       <NA>
#> 7  maritimebnds.54       54          1   49034 Territory sea (12 nm)       <NA>
#> 8  maritimebnds.55       55          1   49036 Territory sea (12 nm)       <NA>
#> 9  maritimebnds.56       56          1   49042 Territory sea (12 nm)       <NA>
#> 10 maritimebnds.57       57          1   49063 Territory sea (12 nm)       <NA>
#>    legalfou_1        country nationalle nutscode               mblsds_mbl
#> 1        <NA>         Latvia       <NA>       LV In www.marineregions.org
#> 2        <NA>         France       <NA>       FR In www.marineregions.org
#> 3        <NA>         France       <NA>       FR In www.marineregions.org
#> 4        <NA>         France       <NA>       FR In www.marineregions.org
#> 5        <NA>          Spain       <NA>       ES In www.marineregions.org
#> 6        <NA>         Cyprus       <NA>       CY In www.marineregions.org
#> 7        <NA>         Sweden       <NA>       SE In www.marineregions.org
#> 8        <NA> United Kingdom       <NA>       UK In www.marineregions.org
#> 9        <NA>         France       <NA>       FR In www.marineregions.org
#> 10       <NA>         France       <NA>       FR In www.marineregions.org
#>    shape_leng                       the_geom
#> 1   12.005746 MULTICURVE (LINESTRING (23....
#> 2    1.949563 MULTICURVE (LINESTRING (42....
#> 3    2.072975 MULTICURVE (LINESTRING (39....
#> 4    1.837879 MULTICURVE (LINESTRING (40....
#> 5   27.878496 MULTICURVE (LINESTRING (-17...
#> 6   13.642806 MULTICURVE (LINESTRING (34....
#> 7   68.013014 MULTICURVE (LINESTRING (19....
#> 8  141.545470 MULTICURVE (LINESTRING (-13...
#> 9   70.017156 MULTICURVE (LINESTRING (169...
#> 10   5.116398 MULTICURVE (LINESTRING (55....