Get EMODnet WFS layers

Get EMODnet WFS layers

emodnet_get_layers(
  wfs = NULL,
  service = "seabed_habitats_individual_habitat_map_and_model_datasets",
  service_version = "2.0.0",
  layers,
  crs = 4326,
  reduce_layers = FALSE,
  suppress_warnings = FALSE
)

Arguments

wfs
A `WFSClient` R6 object with methods for interfacing an OGC Web Feature Service.

service
the EMODnet OGC WFS service name. Defaults to the EMODnet Seabed Habitats WFS `"seabed_habitats_individual_habitat_map_and_model_datasets"`. For available services, see `emodnet_wfs`.

service_version
the WFS service version. Defaults to "2.0.0".

layers
a character vector of layer names. To get info on layers, including `layer_name` use `emodnet_get_wfs_info()`

crs
integer. An EPSG code for the outpur crs. Defaults to 4326, corresponsing to `"+proj=longlat +datum=WGS84 +no_defs"`

reduce_layers
whether to reduce output layers to a single `sf` object.

suppress_warnings
logical. Whether to suppress messages of layer download failures.

Value

If `reduce_layers = FALSE` (default), a list of `sf` objects, one element for each layer. Any layers for which download was unsuccessful will be NULL. If `reduce_layers = TRUE`, all layers are reduced (if possible) to a single `sf` containing data for all layers. `NULL` layers are ignored.

Examples

emodnet_get_layers(layers = c("dk003069", "dk003070"))

#> Loading ISO 19139 XML schemas...

#> Loading ISO 19115 codelists...

#> ✔ WFS client created succesfully

#> ℹ Service: 'https://ows.emodnet-seabedhabitats.eu/emodnet_open_maplibrary/wfs'

#> ℹ Version: '2.0.0'

#> ℹ crs transformed from 3857 to 4326

#> ℹ crs transformed from 3857 to 4326

#> $dk003069
#> Simple feature collection with 82 features and 8 fields
#> Geometry type: MULTISURFACE
#> Dimension:     XY
#> Bounding box:  xmin: 9.575308 ymin: 54.77378 xmax: 10.24418 ymax: 55.12132
#> Geodetic CRS:  WGS 84
#> First 10 features:
#>            gml_id   gid      gui polygon annexi         subtype confidence
#> 1  dk003069.39813 39813 DK003069      30   1170 Geogenic origin       High
#> 2  dk003069.39847 39847 DK003069      64   1110            <NA>       High
#> 3  dk003069.39794 39794 DK003069      11   1170 Geogenic origin       High
#> 4  dk003069.39787 39787 DK003069       4   1170 Geogenic origin       High
#> 5  dk003069.39809 39809 DK003069      26   1170 Geogenic origin       High
#> 6  dk003069.39806 39806 DK003069      23   1170 Geogenic origin       High
#> 7  dk003069.39801 39801 DK003069      18   1170 Geogenic origin       High
#> 8  dk003069.39805 39805 DK003069      22   1170 Geogenic origin       High
#> 9  dk003069.39827 39827 DK003069      44   1170 Geogenic origin       High
#> 10 dk003069.39849 39849 DK003069      66   1110            <NA>       High
#>    val_comm                           geom
#> 1      <NA> MULTISURFACE (POLYGON ((10....
#> 2      <NA> MULTISURFACE (POLYGON ((10....
#> 3      <NA> MULTISURFACE (POLYGON ((10....
#> 4      <NA> MULTISURFACE (POLYGON ((9.8...
#> 5      <NA> MULTISURFACE (POLYGON ((10....
#> 6      <NA> MULTISURFACE (POLYGON ((9.9...
#> 7      <NA> MULTISURFACE (POLYGON ((9.8...
#> 8      <NA> MULTISURFACE (POLYGON ((9.8...
#> 9      <NA> MULTISURFACE (POLYGON ((10....
#> 10     <NA> MULTISURFACE (POLYGON ((9.9...
#> 
#> $dk003070
#> Simple feature collection with 30 features and 8 fields
#> Geometry type: MULTISURFACE
#> Dimension:     XY
#> Bounding box:  xmin: 11.39643 ymin: 54.55514 xmax: 11.96792 ymax: 54.63234
#> Geodetic CRS:  WGS 84
#> First 10 features:
#>            gml_id   gid      gui polygon annexi         subtype confidence
#> 1  dk003070.39886 39886 DK003070      21   1170 Geogenic origin       High
#> 2  dk003070.39890 39890 DK003070      25   1170 Geogenic origin       High
#> 3  dk003070.39882 39882 DK003070      17   1170 Geogenic origin       High
#> 4  dk003070.39875 39875 DK003070      10   1170 Geogenic origin       High
#> 5  dk003070.39881 39881 DK003070      16   1170 Geogenic origin       High
#> 6  dk003070.39887 39887 DK003070      22   1170 Geogenic origin       High
#> 7  dk003070.39885 39885 DK003070      20   1170 Geogenic origin       High
#> 8  dk003070.39879 39879 DK003070      14   1170 Geogenic origin       High
#> 9  dk003070.39867 39867 DK003070       2   1170 Geogenic origin       High
#> 10 dk003070.39873 39873 DK003070       8   1170 Geogenic origin       High
#>    val_comm                           geom
#> 1      <NA> MULTISURFACE (POLYGON ((11....
#> 2      <NA> MULTISURFACE (POLYGON ((11....
#> 3      <NA> MULTISURFACE (POLYGON ((11....
#> 4      <NA> MULTISURFACE (POLYGON ((11....
#> 5      <NA> MULTISURFACE (POLYGON ((11....
#> 6      <NA> MULTISURFACE (POLYGON ((11....
#> 7      <NA> MULTISURFACE (POLYGON ((11....
#> 8      <NA> MULTISURFACE (POLYGON ((11....
#> 9      <NA> MULTISURFACE (POLYGON ((11....
#> 10     <NA> MULTISURFACE (POLYGON ((11....
#> 

emodnet_get_layers(layers = c("dk003069", "dk003070"), reduce_layers = TRUE)

#> ✔ WFS client created succesfully

#> ℹ Service: 'https://ows.emodnet-seabedhabitats.eu/emodnet_open_maplibrary/wfs'

#> ℹ Version: '2.0.0'

#> ℹ crs transformed from 3857 to 4326

#> Simple feature collection with 112 features and 8 fields
#> Geometry type: MULTISURFACE
#> Dimension:     XY
#> Bounding box:  xmin: 9.575308 ymin: 54.55514 xmax: 11.96792 ymax: 55.12132
#> Geodetic CRS:  WGS 84
#> First 10 features:
#>            gml_id   gid      gui polygon annexi         subtype confidence
#> 1  dk003069.39813 39813 DK003069      30   1170 Geogenic origin       High
#> 2  dk003069.39847 39847 DK003069      64   1110            <NA>       High
#> 3  dk003069.39794 39794 DK003069      11   1170 Geogenic origin       High
#> 4  dk003069.39787 39787 DK003069       4   1170 Geogenic origin       High
#> 5  dk003069.39809 39809 DK003069      26   1170 Geogenic origin       High
#> 6  dk003069.39806 39806 DK003069      23   1170 Geogenic origin       High
#> 7  dk003069.39801 39801 DK003069      18   1170 Geogenic origin       High
#> 8  dk003069.39805 39805 DK003069      22   1170 Geogenic origin       High
#> 9  dk003069.39827 39827 DK003069      44   1170 Geogenic origin       High
#> 10 dk003069.39849 39849 DK003069      66   1110            <NA>       High
#>    val_comm                           geom
#> 1      <NA> MULTISURFACE (POLYGON ((10....
#> 2      <NA> MULTISURFACE (POLYGON ((10....
#> 3      <NA> MULTISURFACE (POLYGON ((10....
#> 4      <NA> MULTISURFACE (POLYGON ((9.8...
#> 5      <NA> MULTISURFACE (POLYGON ((10....
#> 6      <NA> MULTISURFACE (POLYGON ((9.9...
#> 7      <NA> MULTISURFACE (POLYGON ((9.8...
#> 8      <NA> MULTISURFACE (POLYGON ((9.8...
#> 9      <NA> MULTISURFACE (POLYGON ((10....
#> 10     <NA> MULTISURFACE (POLYGON ((9.9...