Get WFS available layer information

Get WFS available layer information

Usage

emodnet_get_layer_info(wfs, layers)

emodnet_get_wfs_info(wfs = NULL, service = NULL, service_version = NULL)

emodnet_get_all_wfs_info()

Arguments

wfs

A WFSClient R6 object with methods for interfacing an OGC Web Feature Service.

layers

a character vector of layer names. To get info on layers, including layer_name use emodnet_get_wfs_info().

service

the EMODnet OGC WFS service name. For available services, see emodnet_wfs().

service_version

the WFS service version. Now always "2.0.0".

Value

a tibble containing metadata on each layer available from the service.

Details

To minimize the number of requests sent to webservices, these functions use memoise to cache results inside the active R session. To clear the cache, re-start R or run memoise::forget(emodnet_get_wfs_info)/ memoise::forget(emodnet_get_layer_info).

Functions

emodnet_get_layer_info(): Get metadata for specific layers. Requires a wfs object as input.

emodnet_get_wfs_info(): Get info on all layers from an EMODnet WFS service.

emodnet_get_all_wfs_info(): Get metadata on all layers and all available services from server.

Examples

if (FALSE) {
emodnet_get_wfs_info(service = "bathymetry")
# Query a wfs object
wfs_cml <- emodnet_init_wfs_client("chemistry_marine_litter")
emodnet_get_wfs_info(wfs_cml)
# Get info for specific layers from wfs object
layers <- c(
  "bl_fishing_monitoring",
  "bl_beacheslocations_monitoring"
)
emodnet_get_layer_info(wfs = wfs_cml, layers = layers)
}