Get WFS available layer information

Get WFS available layer information

Usage

emodnet_get_wfs_info(wfs = NULL, service = NULL, service_version = "2.0.0")

emodnet_get_layer_info(wfs, layers)

emodnet_get_all_wfs_info()

Arguments

wfs

A WFSClient R6 object with methods for interfacing an OGC Web Feature Service.

service

the EMODnet OGC WFS service name. For available services, see emodnet_wfs.

service_version

the WFS service version. Defaults to "2.0.0".

layers

a character vector of layer names. To get info on layers, including layer_name use emodnet_get_wfs_info().

Value

a tibble containg metadata on each layer available from the service.

Functions

emodnet_get_wfs_info: Get info on all layers from am EMODnet WFS service.

emodnet_get_layer_info: Get metadata for specific layers. Requires a wfs object as input.

emodnet_get_all_wfs_info: Get metadata on all layers and all available services from server.

Examples

emodnet_get_wfs_info(service = "bathymetry")
#> ✔ WFS client created succesfully
#> ℹ Service: 'https://ows.emodnet-bathymetry.eu/wfs'
#> ℹ Version: '2.0.0'
#> # A tibble: 3 × 9
#>   data_source service_name service_url layer_namespace layer_name title abstract
#>   <chr>       <chr>        <chr>       <chr>           <chr>      <chr> <chr>   
#> 1 emodnet_wfs bathymetry   https://ow… emodnet         contours   Dept… "Genera…
#> 2 emodnet_wfs bathymetry   https://ow… emodnet         quality_i… Qual… "Repres…
#> 3 emodnet_wfs bathymetry   https://ow… emodnet         source_re… Sour… "Covera…
#> # … with 2 more variables: class <chr>, format <chr>
# Query a wfs object
wfs_cml <- emodnet_init_wfs_client("chemistry_marine_litter")
#> ✔ WFS client created succesfully
#> ℹ Service: 'https://www.ifremer.fr/services/wfs/emodnet_chemistry2'
#> ℹ Version: '2.0.0'
emodnet_get_wfs_info(wfs_cml)
#> # A tibble: 24 × 9
#>    data_source service_name         service_url layer_namespace layer_name title
#>    <chr>       <chr>                <chr>       <chr>           <chr>      <chr>
#>  1 emodnet_wfs chemistry_marine_li… https://ww… ms              bl_beache… Beac…
#>  2 emodnet_wfs chemistry_marine_li… https://ww… ms              bl_tempor… Numb…
#>  3 emodnet_wfs chemistry_marine_li… https://ww… ms              bl_totala… Beac…
#>  4 emodnet_wfs chemistry_marine_li… https://ww… ms              bl_tot_ab… Beac…
#>  5 emodnet_wfs chemistry_marine_li… https://ww… ms              bl_materi… Beac…
#>  6 emodnet_wfs chemistry_marine_li… https://ww… ms              bl_cigare… Beac…
#>  7 emodnet_wfs chemistry_marine_li… https://ww… ms              bl_cigare… Beac…
#>  8 emodnet_wfs chemistry_marine_li… https://ww… ms              bl_fishin… Beac…
#>  9 emodnet_wfs chemistry_marine_li… https://ww… ms              bl_plasti… Beac…
#> 10 emodnet_wfs chemistry_marine_li… https://ww… ms              bl_single… Beac…
#> # … with 14 more rows, and 3 more variables: abstract <chr>, class <chr>,
#> #   format <chr>
# Get info for specific layers from wfs object
layers <- c("bl_fishing_monitoring",
           "bl_beacheslocations_monitoring")
emodnet_get_layer_info(wfs = wfs_cml, layers = layers)
#> # A tibble: 2 × 9
#>   data_source service_name service_url layer_namespace layer_name title abstract
#>   <chr>       <chr>        <chr>       <chr>           <chr>      <chr> <chr>   
#> 1 emodnet_wfs https://www… chemistry_… ms              bl_beache… Beac… ""      
#> 2 emodnet_wfs https://www… chemistry_… ms              bl_fishin… Beac… ""      
#> # … with 2 more variables: class <chr>, format <chr>